Fix some more compiler warnings. Re-enable detection of unused code.

+++ arch/dev/RPC_unix.cpp
@@ -79,12 +79,12 @@ namespace Simulator
 
 #define RequireArgs(Arg1Size, Arg2Size)                                 \
     do {                                                                \
-        if ((Arg1Size) && arg1.size() < (Arg1Size))                     \
+        if ((Arg1Size) && (int)arg1.size() < (Arg1Size))                \
             throw exceptf<SimulationException>("Procedure %u requires at least %zu bytes in 1st memory argument, %zu were provided", \
                                                (unsigned)procedure_id,  \
                                                (size_t)(Arg1Size),      \
                                                arg1.size());            \
-        if ((Arg2Size) && arg2.size() < (Arg2Size))                     \
+        if ((Arg2Size) && (int)arg2.size() < (Arg2Size))                \
             throw exceptf<SimulationException>("Procedure %u requires at least %zu bytes in 2nd memory argument, %zu were provided", \
                                                (unsigned)procedure_id,  \
                                                (size_t)(Arg2Size),      \
